% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{county_pop}
\alias{county_pop}
\title{County level population 1749-2010}
\format{A data frame with 1358 rows and 4 variables:
\describe{
  \item{county}{County code}
  \item{name}{Name of county}
  \item{year}{Year}
  \item{pop}{Population}
}}
\source{
Statistiska centralbyrån.
}
\usage{
data(county_pop)
}
\description{
County level population 1749-2010.
}

