% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{imr}
\alias{imr}
\title{Infant mortality rates}
\format{A data frame with 550 rows and 4 variables:
\describe{
  \item{county}{County code}
  \item{imr}{Infant mortality rate}
  \item{from}{Start year}
  \item{to}{End year}
}}
\source{
1811-1859, Tabellverket. 1860-1866
  from Th. Berg, Om dödligheten i första lefnadsåret, Statistisk
  tidskrift 1869. 1871-1952, Hofsten Lundström Swedish Population History
}
\usage{
imr
}
\description{
County level inafant mortality rates 1811-1972
}

