% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{mfrt}
\alias{mfrt}
\title{MFRT in Sweden 1751-1970}
\format{A data frame with 1678 rows and 6 variables:
\describe{
  \item{from}{Start year}
  \item{to}{End year}
  \item{context}{Context name, county or Sweden total}
  \item{age}{Age group}
  \item{fert}{Fertility rate}
  \item{county_code}{County code}
}}
\source{
Table 2.3 Marital fertility rates for period, 1751-1970;
  Table 6.18 Age specific rates of marital fertlity for the counties
  1881-1962. Hofsten, E. & Lundström, H. (1976). Swedish population
  history: main trends from 1750 to 1970. [Stockholm]: [LiberFörlag].
}
\usage{
data(mfrt)
}
\description{
Martial fertility rates in Sweden 1751 to 1970 by county.
}

