% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{pop_fert}
\alias{pop_fert}
\title{Age-specific Fertility in Sweden 1751-1967}
\format{A data frame with 1890 rows and 5 variables:
\describe{
  \item{context}{Context area, county name or national level}
  \item{county_code}{County code}
  \item{start}{Start year}
  \item{end}{End year}
  \item{age_g}{Age group}
  \item{fert}{Fertility rate}
}}
\source{
Hofsten, E. & Lundström, H. (1976). Swedish population
  history: main trends from 1750 to 1970. [Stockholm]: [LiberFörlag].
}
\usage{
data(pop_fert)
}
\description{
National level age-specific fertility rates in Sweden
  1751 to 1967.
}

