% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{swe_pop_age}
\alias{swe_pop_age}
\title{National population by age 1750-1967}
\format{A data frame with 300 rows and 3 variables:
\describe{
  \item{year}{Year}
  \item{age}{Age group}
  \item{n}{Population in group}
}}
\source{
Sverige. Statistiska centralbyrån (1969). Historisk statistik för
  Sverige. D. 1, Befolkning 1720-1967 = [Population 1720-1967]. (2. uppl.)
  Örebro: Statistiska centralbyrån (SCB).
}
\usage{
data(swe_pop_age)
}
\description{
Population by age 1750-1967 in Sweden on a national level
}

