% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{pop_age}
\alias{pop_age}
\title{National population by age and gender 1750-2014}
\format{A data frame with 31240 rows and 5 variables:
\describe{
  \item{year}{Year}
  \item{gender}{Male, female or all}
  \item{pop}{Population}
  \item{age}{Star age of age group}
  \item{span}{Age span of age group}
}}
\source{
Statistiska centralbyrån. \url{http://www.scb.se/en_/Finding-statistics/
  Statistics-by-subject-area/Population/Population-composition/Population-statistics/
  Aktuell-Pong/25795/Yearly-statistics--The-whole-country/105505/};
  Hofsten, E. & Lundström, H. (1976). Swedish population
  history: main trends from 1750 to 1970. [Stockholm]: [LiberFörlag].
}
\usage{
data(pop_age)
}
\description{
Population by age and gender 1750-2014 in Sweden on a national level
}

